requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --no-annotate
#
aioredis==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
automat==22.10.0
brotli==1.0.9
cachecontrol==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==39.0.2
daphne==3.0.2
dj-database-url==1.2.0
django==3.2.18
djangorestframework==3.14.0
firebase-admin==6.1.0
google-api-core[grpc]==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
hiredis==2.2.2
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
phonenumbers==8.13.10
proto-plus==1.22.2
protobuf==4.22.1
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyjwt[crypto]==2.6.0
pyopenssl==23.0.0
pyparsing==3.0.9
pytz==2022.7.1
requests==2.28.2
rsa==4.9
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
twisted[tls]==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
whitenoise[brotli]==6.5.0
zope-interface==5.5.2

# The following packages are considered to be unsafe in a requirements file:
# setuptools
